requirements.txt
backcall==0.1.0
beautifulsoup4==4.6.0
certifi==2018.8.24
chardet==3.0.4
decorator==4.3.0
Django==2.0.8
django-activeurl==0.1.12
django-appconf==1.0.2
django-classy-tags==0.8.0
django-columns==0.1.0
django-extensions==2.1.3
django-modelcluster==4.2
django-taggit==0.23.0
django-treebeard==4.3
djangorestframework==3.8.2
draftjs-exporter==2.1.2
html5lib==1.0.1
idna==2.7
ipython==7.0.1
ipython-genutils==0.2.0
jedi==0.13.1
lxml==4.2.5
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.2.0
prompt-toolkit==2.0.5
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
Pygments==2.2.0
python-decouple==3.1
pytils==0.3
pytz==2018.5
requests==2.19.1
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
Unidecode==1.0.22
urllib3==1.23
wagtail==2.3
wcwidth==0.1.7
webencodings==0.5.1
Willow==1.1
